Optimized use of GPU memory
Added "number_of_times_to_upsample=0" to face_locations, this will remove the need to resize the frame and increases the amount of faces found.
Increased maximum frame size to 1080x1920, this comes to a GPU memory usage of 3.6 GB for 24-bit color depth. 
Succesfully tested on a HD episode of Game of Thrones with a 4GB NVIDIA GTX 1050Ti without any "cudaMalloc" errors.